annotating useless code
