+++ pkg/microservice/aslan/core/common/service/render.go
@@ -48,11 +48,11 @@ type KVPair struct {
 	Value string `json:"value"`
 }
 
-type YamlData struct {
-	YamlSource    string      `json:"yamlSource,omitempty"`
-	GitRepoConfig *RepoConfig `json:"gitRepoConfig,omitempty"`
-	ValuesYAML    string      `json:"valuesYAML,omitempty"`
-}
+//type YamlData struct {
+//YamlSource    string      `json:"yamlSource,omitempty"`
+//GitRepoConfig *RepoConfig `json:"gitRepoConfig,omitempty"`
+//ValuesYAML string `json:"valuesYAML,omitempty"`
+//}
 
 type RenderChartArg struct {
 	EnvName        string    `json:"envName,omitempty"`
@@ -60,13 +60,12 @@ type RenderChartArg struct {
 	ChartVersion   string    `json:"chartVersion,omitempty"`
 	OverrideValues []*KVPair `json:"overrideValues,omitempty"`
 	OverrideYaml   string    `json:"overrideYaml,omitempty"`
-	//*YamlData
 }
 
-type RendersetArg struct {
-	DefaultValues *YamlData `json:"defaultValues"`
-	*RenderChartArg
-}
+//type RendersetArg struct {
+//	DefaultValues *YamlData `json:"defaultValues"`
+//	*RenderChartArg
+//}
 
 func (args *RenderChartArg) ToOverrideValueString() string {
 	if len(args.OverrideValues) == 0 {
