+++ pkg/microservice/aslan/core/environment/handler/router.go
@@ -135,7 +135,7 @@ func (*Router) Inject(router *gin.RouterGroup) {
 		environments.GET("/estimated-renderchart", GetEstimatedRenderCharts)
 
 		//TODO need to be deprecated
-		environments.PUT("/:productName/renderchart", gin2.IsHavePermission([]string{permission.TestEnvManageUUID}, permission.ParamType), gin2.UpdateOperationLogStatus, UpdateHelmProductRenderCharts)
+		//environments.PUT("/:productName/renderchart", gin2.IsHavePermission([]string{permission.TestEnvManageUUID}, permission.ParamType), gin2.UpdateOperationLogStatus, UpdateHelmProductRenderCharts)
 	}
 
 	// ---------------------------------------------------------------------------------------
@@ -148,9 +148,9 @@ func (*Router) Inject(router *gin.RouterGroup) {
 		rendersets.GET("/yamlContent", GetYamlContent)
 
 		//TODO  need to deprecated
-		rendersets.PUT("/renderset", CreateOrUpdateRenderset)
+		//rendersets.PUT("/renderset", CreateOrUpdateRenderset)
 		//TODO  need to deprecated
-		rendersets.PUT("/renderchart", CreateOrUpdateRenderChart) // create or update renderchart when onBoarding
+		//rendersets.PUT("/renderchart", CreateOrUpdateRenderChart) // create or update renderchart when onBoarding
 	}
 
 	// ---------------------------------------------------------------------------------------
