+++ pkg/microservice/aslan/core/environment/service/renderset.go
@@ -19,21 +19,15 @@ package service
 import (
 	"fmt"
 	"strings"
-	"sync"
 
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
-	"gopkg.in/yaml.v3"
 	"k8s.io/apimachinery/pkg/util/sets"
 
 	commonmodels "github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/models"
 	"github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/models/template"
 	commonrepo "github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/mongodb"
 	commonservice "github.com/koderover/zadig/pkg/microservice/aslan/core/common/service"
-	fsservice "github.com/koderover/zadig/pkg/microservice/aslan/core/common/service/fs"
-	"github.com/koderover/zadig/pkg/setting"
-	e "github.com/koderover/zadig/pkg/tool/errors"
-	yamlutil "github.com/koderover/zadig/pkg/util/yaml"
 )
 
 type DefaultValuesResp struct {
@@ -102,222 +96,222 @@ func GetDefaultValues(productName, envName string, log *zap.SugaredLogger) (*Def
 }
 
 // validate yaml content
-func validateYamlContent(yamlContent string) error {
-	tMap := map[string]interface{}{}
-	if err := yaml.Unmarshal([]byte(yamlContent), &tMap); err != nil {
-		if err != nil {
-			return errors.New("yaml content illegal")
-		}
-	}
-	return nil
-}
+//func validateYamlContent(yamlContent string) error {
+//	tMap := map[string]interface{}{}
+//	if err := yaml.Unmarshal([]byte(yamlContent), &tMap); err != nil {
+//		if err != nil {
+//			return errors.New("yaml content illegal")
+//		}
+//	}
+//	return nil
+//}
 
-func generateValuesYaml(yamlData *commonservice.YamlData, log *zap.SugaredLogger) error {
-	if yamlData == nil {
-		return nil
-	}
-	if yamlData.YamlSource == setting.ValuesYamlSourceFreeEdit {
-		return validateYamlContent(yamlData.ValuesYAML)
-	} else if yamlData.YamlSource == setting.ValuesYamlSourceGitRepo {
-		if yamlData.GitRepoConfig == nil {
-			return errors.New("invalid repo config")
-		}
-		var (
-			allValues      []byte
-			fileContentMap sync.Map
-			wg             sync.WaitGroup
-			err            error
-		)
+//func generateValuesYaml(yamlData *commonservice.YamlData, log *zap.SugaredLogger) error {
+//	if yamlData == nil {
+//		return nil
+//	}
+//	if yamlData.YamlSource == setting.ValuesYamlSourceFreeEdit {
+//		return validateYamlContent(yamlData.ValuesYAML)
+//	} else if yamlData.YamlSource == setting.ValuesYamlSourceGitRepo {
+//		if yamlData.GitRepoConfig == nil {
+//			return errors.New("invalid repo config")
+//		}
+//		var (
+//			allValues      []byte
+//			fileContentMap sync.Map
+//			wg             sync.WaitGroup
+//			err            error
+//		)
+//
+//		for index, filePath := range yamlData.GitRepoConfig.ValuesPaths {
+//			wg.Add(1)
+//			go func(index int, path string) {
+//				defer wg.Done()
+//				fileContent, err1 := fsservice.DownloadFileFromSource(
+//					&fsservice.DownloadFromSourceArgs{
+//						CodehostID: yamlData.GitRepoConfig.CodehostID,
+//						Owner:      yamlData.GitRepoConfig.Owner,
+//						Repo:       yamlData.GitRepoConfig.Repo,
+//						Path:       path,
+//						Branch:     yamlData.GitRepoConfig.Branch,
+//					})
+//				if err1 != nil {
+//					err = errors.Errorf("fail to download file from git, err: %s, path: %s, repo: %v", err1.Error(), path, *yamlData.GitRepoConfig)
+//					return
+//				}
+//				fileContentMap.Store(index, fileContent)
+//			}(index, filePath)
+//		}
+//		wg.Wait()
+//
+//		if err != nil {
+//			return err
+//		}
+//
+//		allValueYamls := make([][]byte, len(yamlData.GitRepoConfig.ValuesPaths), len(yamlData.GitRepoConfig.ValuesPaths))
+//		for i := 0; i < len(yamlData.GitRepoConfig.ValuesPaths); i++ {
+//			contentObj, _ := fileContentMap.Load(i)
+//			allValueYamls[i] = contentObj.([]byte)
+//		}
+//		allValues, err = yamlutil.Merge(allValueYamls)
+//		if err != nil {
+//			return errors.Errorf("failed to merge yaml files, repo: %v", *yamlData.GitRepoConfig)
+//		}
+//		yamlData.ValuesYAML = string(allValues)
+//	}
+//	return nil
+//}
 
-		for index, filePath := range yamlData.GitRepoConfig.ValuesPaths {
-			wg.Add(1)
-			go func(index int, path string) {
-				defer wg.Done()
-				fileContent, err1 := fsservice.DownloadFileFromSource(
-					&fsservice.DownloadFromSourceArgs{
-						CodehostID: yamlData.GitRepoConfig.CodehostID,
-						Owner:      yamlData.GitRepoConfig.Owner,
-						Repo:       yamlData.GitRepoConfig.Repo,
-						Path:       path,
-						Branch:     yamlData.GitRepoConfig.Branch,
-					})
-				if err1 != nil {
-					err = errors.Errorf("fail to download file from git, err: %s, path: %s, repo: %v", err1.Error(), path, *yamlData.GitRepoConfig)
-					return
-				}
-				fileContentMap.Store(index, fileContent)
-			}(index, filePath)
-		}
-		wg.Wait()
+//func CreateOrUpdateRenderset(productName, envName string, args *commonservice.RendersetArg, userName, requestID string, log *zap.SugaredLogger) error {
+//	serviceName := args.ServiceName
+//
+//	serviceOpt := &commonrepo.ServiceFindOption{
+//		ProductName: productName,
+//		ServiceName: serviceName,
+//		Type:        setting.HelmDeployType,
+//	}
+//	serviceObj, err := commonrepo.NewServiceColl().Find(serviceOpt)
+//	if err != nil {
+//		return e.ErrCreateRenderSet.AddDesc(fmt.Sprintf("failed to get service %s", err.Error()))
+//	}
+//	if serviceObj == nil {
+//		return e.ErrCreateRenderSet.AddDesc("service not found")
+//	}
+//
+//	if serviceObj.HelmChart == nil {
+//		return e.ErrCreateRenderSet.AddDesc("missing helm chart info")
+//	}
+//
+//	if args.DefaultValues != nil && args.DefaultValues.YamlSource != "" {
+//		err := generateValuesYaml(args.DefaultValues, log)
+//		if err != nil {
+//			return e.ErrCreateRenderSet.AddDesc(err.Error())
+//		}
+//	}
+//
+//	renderSetName := commonservice.GetProductEnvNamespace(envName, productName, "")
+//
+//	opt := &commonrepo.RenderSetFindOption{Name: renderSetName}
+//	curRenderset, found, err := commonrepo.NewRenderSetColl().FindRenderSet(opt)
+//	if err != nil {
+//		return e.ErrCreateRenderSet.AddDesc(fmt.Sprintf("failed to get renderset %s", err.Error()))
+//	}
+//
+//	if !found {
+//		curRenderset = &commonmodels.RenderSet{
+//			Name:        renderSetName,
+//			EnvName:     envName,
+//			ProductTmpl: productName,
+//			UpdateBy:    userName,
+//			IsDefault:   false,
+//		}
+//	}
+//
+//	//update or insert service values.yaml
+//	if args.RenderChartArg != nil {
+//		var targetChartInfo *template.RenderChart
+//		for _, singleChart := range curRenderset.ChartInfos {
+//			if singleChart.ServiceName != serviceName {
+//				continue
+//			}
+//			args.RenderChartArg.FillRenderChartModel(singleChart, singleChart.ChartVersion)
+//			targetChartInfo = singleChart
+//			break
+//		}
+//		if targetChartInfo == nil {
+//			targetChartInfo = new(template.RenderChart)
+//			targetChartInfo.ValuesYaml = serviceObj.HelmChart.ValuesYaml
+//			args.FillRenderChartModel(targetChartInfo, serviceObj.HelmChart.Version)
+//			curRenderset.ChartInfos = append(curRenderset.ChartInfos, targetChartInfo)
+//		}
+//	}
+//
+//	//update environment's defaults.yaml
+//	if args.DefaultValues != nil {
+//		curRenderset.DefaultValues = args.DefaultValues.ValuesYAML
+//	} else {
+//		curRenderset.DefaultValues = ""
+//	}
+//
+//	//create new renderset with increased revision
+//	err = commonservice.CreateHelmRenderSet(
+//		curRenderset,
+//		log,
+//	)
+//
+//	if err != nil {
+//		return e.ErrCreateRenderSet.AddDesc(err.Error())
+//	}
+//	return err
+//}
 
-		if err != nil {
-			return err
-		}
-
-		allValueYamls := make([][]byte, len(yamlData.GitRepoConfig.ValuesPaths), len(yamlData.GitRepoConfig.ValuesPaths))
-		for i := 0; i < len(yamlData.GitRepoConfig.ValuesPaths); i++ {
-			contentObj, _ := fileContentMap.Load(i)
-			allValueYamls[i] = contentObj.([]byte)
-		}
-		allValues, err = yamlutil.Merge(allValueYamls)
-		if err != nil {
-			return errors.Errorf("failed to merge yaml files, repo: %v", *yamlData.GitRepoConfig)
-		}
-		yamlData.ValuesYAML = string(allValues)
-	}
-	return nil
-}
-
-func CreateOrUpdateRenderset(productName, envName string, args *commonservice.RendersetArg, userName, requestID string, log *zap.SugaredLogger) error {
-	serviceName := args.ServiceName
-
-	serviceOpt := &commonrepo.ServiceFindOption{
-		ProductName: productName,
-		ServiceName: serviceName,
-		Type:        setting.HelmDeployType,
-	}
-	serviceObj, err := commonrepo.NewServiceColl().Find(serviceOpt)
-	if err != nil {
-		return e.ErrCreateRenderSet.AddDesc(fmt.Sprintf("failed to get service %s", err.Error()))
-	}
-	if serviceObj == nil {
-		return e.ErrCreateRenderSet.AddDesc("service not found")
-	}
-
-	if serviceObj.HelmChart == nil {
-		return e.ErrCreateRenderSet.AddDesc("missing helm chart info")
-	}
-
-	if args.DefaultValues != nil && args.DefaultValues.YamlSource != "" {
-		err := generateValuesYaml(args.DefaultValues, log)
-		if err != nil {
-			return e.ErrCreateRenderSet.AddDesc(err.Error())
-		}
-	}
-
-	renderSetName := commonservice.GetProductEnvNamespace(envName, productName, "")
-
-	opt := &commonrepo.RenderSetFindOption{Name: renderSetName}
-	curRenderset, found, err := commonrepo.NewRenderSetColl().FindRenderSet(opt)
-	if err != nil {
-		return e.ErrCreateRenderSet.AddDesc(fmt.Sprintf("failed to get renderset %s", err.Error()))
-	}
-
-	if !found {
-		curRenderset = &commonmodels.RenderSet{
-			Name:        renderSetName,
-			EnvName:     envName,
-			ProductTmpl: productName,
-			UpdateBy:    userName,
-			IsDefault:   false,
-		}
-	}
-
-	//update or insert service values.yaml
-	if args.RenderChartArg != nil {
-		var targetChartInfo *template.RenderChart
-		for _, singleChart := range curRenderset.ChartInfos {
-			if singleChart.ServiceName != serviceName {
-				continue
-			}
-			args.RenderChartArg.FillRenderChartModel(singleChart, singleChart.ChartVersion)
-			targetChartInfo = singleChart
-			break
-		}
-		if targetChartInfo == nil {
-			targetChartInfo = new(template.RenderChart)
-			targetChartInfo.ValuesYaml = serviceObj.HelmChart.ValuesYaml
-			args.FillRenderChartModel(targetChartInfo, serviceObj.HelmChart.Version)
-			curRenderset.ChartInfos = append(curRenderset.ChartInfos, targetChartInfo)
-		}
-	}
-
-	//update environment's defaults.yaml
-	if args.DefaultValues != nil {
-		curRenderset.DefaultValues = args.DefaultValues.ValuesYAML
-	} else {
-		curRenderset.DefaultValues = ""
-	}
-
-	//create new renderset with increased revision
-	err = commonservice.CreateHelmRenderSet(
-		curRenderset,
-		log,
-	)
-
-	if err != nil {
-		return e.ErrCreateRenderSet.AddDesc(err.Error())
-	}
-	return err
-}
-
-func CreateOrUpdateChartValues(productName, envName string, args *commonservice.RenderChartArg, userName, requestID string, log *zap.SugaredLogger) error {
-
-	serviceName := args.ServiceName
-
-	serviceOpt := &commonrepo.ServiceFindOption{
-		ProductName: productName,
-		ServiceName: serviceName,
-		Type:        setting.HelmDeployType,
-	}
-	serviceObj, err := commonrepo.NewServiceColl().Find(serviceOpt)
-	if err != nil {
-		return e.ErrCreateRenderSet.AddDesc(fmt.Sprintf("failed to get service %s", err.Error()))
-	}
-	if serviceObj == nil {
-		return e.ErrCreateRenderSet.AddDesc("service not found")
-	}
-
-	if serviceObj.HelmChart == nil {
-		return e.ErrCreateRenderSet.AddDesc("missing helm chart info")
-	}
-
-	renderSetName := commonservice.GetProductEnvNamespace(envName, productName, "")
-
-	opt := &commonrepo.RenderSetFindOption{Name: renderSetName}
-	curRenderset, found, err := commonrepo.NewRenderSetColl().FindRenderSet(opt)
-	if err != nil {
-		return e.ErrCreateRenderSet.AddDesc(fmt.Sprintf("failed to get renderset %s", err.Error()))
-	}
-
-	if !found {
-		curRenderset = &commonmodels.RenderSet{
-			Name:        renderSetName,
-			EnvName:     envName,
-			ProductTmpl: productName,
-			UpdateBy:    userName,
-			IsDefault:   false,
-		}
-	}
-
-	//update or insert service values.yaml
-	var targetChartInfo *template.RenderChart
-	for _, singleChart := range curRenderset.ChartInfos {
-		if singleChart.ServiceName != serviceName {
-			continue
-		}
-		args.FillRenderChartModel(singleChart, singleChart.ChartVersion)
-		targetChartInfo = singleChart
-		break
-	}
-	if targetChartInfo == nil {
-		targetChartInfo = new(template.RenderChart)
-		targetChartInfo.ValuesYaml = serviceObj.HelmChart.ValuesYaml
-		args.FillRenderChartModel(targetChartInfo, serviceObj.HelmChart.Version)
-		curRenderset.ChartInfos = append(curRenderset.ChartInfos, targetChartInfo)
-	}
-
-	//create new renderset with increased revision
-	err = commonservice.CreateHelmRenderSet(
-		curRenderset,
-		log,
-	)
-
-	if err != nil {
-		return e.ErrCreateRenderSet.AddDesc(err.Error())
-	}
-	return err
-}
+//func CreateOrUpdateChartValues(productName, envName string, args *commonservice.RenderChartArg, userName, requestID string, log *zap.SugaredLogger) error {
+//
+//	serviceName := args.ServiceName
+//
+//	serviceOpt := &commonrepo.ServiceFindOption{
+//		ProductName: productName,
+//		ServiceName: serviceName,
+//		Type:        setting.HelmDeployType,
+//	}
+//	serviceObj, err := commonrepo.NewServiceColl().Find(serviceOpt)
+//	if err != nil {
+//		return e.ErrCreateRenderSet.AddDesc(fmt.Sprintf("failed to get service %s", err.Error()))
+//	}
+//	if serviceObj == nil {
+//		return e.ErrCreateRenderSet.AddDesc("service not found")
+//	}
+//
+//	if serviceObj.HelmChart == nil {
+//		return e.ErrCreateRenderSet.AddDesc("missing helm chart info")
+//	}
+//
+//	renderSetName := commonservice.GetProductEnvNamespace(envName, productName, "")
+//
+//	opt := &commonrepo.RenderSetFindOption{Name: renderSetName}
+//	curRenderset, found, err := commonrepo.NewRenderSetColl().FindRenderSet(opt)
+//	if err != nil {
+//		return e.ErrCreateRenderSet.AddDesc(fmt.Sprintf("failed to get renderset %s", err.Error()))
+//	}
+//
+//	if !found {
+//		curRenderset = &commonmodels.RenderSet{
+//			Name:        renderSetName,
+//			EnvName:     envName,
+//			ProductTmpl: productName,
+//			UpdateBy:    userName,
+//			IsDefault:   false,
+//		}
+//	}
+//
+//	//update or insert service values.yaml
+//	var targetChartInfo *template.RenderChart
+//	for _, singleChart := range curRenderset.ChartInfos {
+//		if singleChart.ServiceName != serviceName {
+//			continue
+//		}
+//		args.FillRenderChartModel(singleChart, singleChart.ChartVersion)
+//		targetChartInfo = singleChart
+//		break
+//	}
+//	if targetChartInfo == nil {
+//		targetChartInfo = new(template.RenderChart)
+//		targetChartInfo.ValuesYaml = serviceObj.HelmChart.ValuesYaml
+//		args.FillRenderChartModel(targetChartInfo, serviceObj.HelmChart.Version)
+//		curRenderset.ChartInfos = append(curRenderset.ChartInfos, targetChartInfo)
+//	}
+//
+//	//create new renderset with increased revision
+//	err = commonservice.CreateHelmRenderSet(
+//		curRenderset,
+//		log,
+//	)
+//
+//	if err != nil {
+//		return e.ErrCreateRenderSet.AddDesc(err.Error())
+//	}
+//	return err
+//}
 
 func ensureHelmRenderSetArgs(args *commonmodels.RenderSet) error {
 	if args == nil {
